% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLSVDS_calc.R
\name{All_SVDS}
\alias{All_SVDS}
\title{Aggregate SVD Results from Multiple Models}
\usage{
All_SVDS(
  data_not_contaminated,
  data_contaminated,
  weights,
  Ncomp = 2,
  rSVD = FALSE
)
}
\arguments{
\item{data_not_contaminated}{The original (non-contaminated) data frame.}

\item{data_contaminated}{The contaminated data frame.}

\item{weights}{A list of weight vectors for the models.}

\item{Ncomp}{Number of SVD components to retain. Default is 2.}

\item{rSVD}{Logical; if \code{TRUE}, use robust SVD in RW-AMMI. Default is \code{FALSE}.}
}
\value{
A list containing the fitted models for:
  \describe{
    \item{RealSVD}{AMMI model fitted to the original data.}
    \item{AMMI}{AMMI model fitted to the contaminated data.}
    \item{WAMMI}{List of W-AMMI models for different weight schemes.}
    \item{RAMMI}{List of R-AMMI models for different weight schemes.}
    \item{RWAMMI}{List of RW-AMMI models for different weight schemes.}
  }
}
\description{
This function fits several models (AMMI, W-AMMI, R-AMMI, and RW-AMMI) to the provided data,
and aggregates their singular value decomposition (SVD) results.
}
\examples{
sim_data <- sim.amb(seed = 123)
rep_data <- Create.Replications(sim_data, reps = 2, sig = 1)
contaminated <- Data_Contamination(rep_data, percentage = 5, seed = 1, type = "shift")
# Assume weights is a predefined list of weight vectors
models <- All_SVDS(data_not_contaminated = rep_data,
                   data_contaminated = contaminated$data.contaminated,
                   weights = list(LMM = list(w1 = rep(1, nrow(rep_data))),
                                  RLMM = list(w1 = rep(1, nrow(rep_data)))),
                   Ncomp = 2, rSVD = FALSE)
names(models)

}
