% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMMI_Model.R
\name{ammi.model}
\alias{ammi.model}
\title{Fit an AMMI Model}
\usage{
ammi.model(dataframe, Ncomp = 2)
}
\arguments{
\item{dataframe}{A data frame with genotype-environment data.}

\item{Ncomp}{Number of components to retain in the SVD. Default is 2.}
}
\value{
A list containing:
  \describe{
    \item{estimated.data}{The fitted data including main and multiplicative effects.}
    \item{anovatable}{The ANOVA table of the fitted linear model.}
    \item{SVD}{The SVD object from the residual matrix.}
    \item{residual}{The residual matrix from the fitted model.}
    \item{matrix.fitted}{The matrix of fitted values from the linear model.}
    \item{reduced.SVD}{The SVD reconstruction using only the first \code{Ncomp} components.}
  }
}
\description{
This function fits the Additive Main effects and Multiplicative Interaction (AMMI) model.
It calculates fitted values, residuals, and performs singular value decomposition on the residuals.
}
\examples{
sim_data <- sim.amb(seed = 123)
ammi_res <- ammi.model(sim_data, Ncomp = 2)
names(ammi_res)

}
