% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replications.R
\name{Create.Replications}
\alias{Create.Replications}
\title{Create Replications of Genotype-Environment Data}
\usage{
Create.Replications(data, reps = 2, sig = 1)
}
\arguments{
\item{data}{A data frame containing the original genotype-environment data. The data frame must have the following columns:
\describe{
  \item{gen}{A factor representing genotype identifiers.}
  \item{env}{A factor representing environment identifiers.}
  \item{yield}{A numeric vector representing the response variable.}
}}

\item{reps}{An integer specifying the number of replications to create. Default is 2.}

\item{sig}{A numeric value specifying the standard deviation of the residuals to add. Default is 1.}
}
\value{
A data frame with the replicated data. Each replication is identified by the "rep" column.
}
\description{
This function generates replications of a genotype-environment dataset by adding random residuals.
It simulates experimental replications by perturbing the response variable (yield) with residual error,
allowing for the assessment of variability across replications.
}
\examples{
# Generate simulated data using sim.amb (assumes sim.amb is already defined and documented)
sim_data <- sim.amb(seed = 123)

# Create 3 replications with a residual standard deviation of 1.5
rep_data <- Create.Replications(sim_data, reps = 3, sig = 1.5)
head(rep_data)

}
