% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contamination_data.R
\name{Data_Contamination}
\alias{Data_Contamination}
\title{Contaminate Genotype-Environment Data}
\usage{
Data_Contamination(
  data,
  percentage = 5,
  seed = 1,
  type = "shift",
  k = 7,
  c = 10
)
}
\arguments{
\item{data}{A data frame containing the original genotype-environment data.}

\item{percentage}{Percentage of data to contaminate. If greater than 1, it is treated as a percentage.
Default is 5.}

\item{seed}{Seed value for reproducibility. Default is 1.}

\item{type}{Type of contamination. Must be one of \code{"shift"}, \code{"pointmass"}, or \code{"varinflated"}.
Default is \code{"shift"}.}

\item{k}{Shift factor for contamination if \code{type} is "shift". Default is 7.}

\item{c}{Scaling factor for contamination if \code{type} is "pointmass" or "varinflated". Default is 10.}
}
\value{
A list with two elements:
  \describe{
    \item{data.not_contaminated}{The original data.}
    \item{data.contaminated}{The contaminated dataset.}
  }
}
\description{
This function introduces contamination into the first replication of a genotype-environment dataset.
Different types of contamination can be applied: shifting values, creating point mass contamination,
or inflating variances.
}
\examples{
sim_data <- sim.amb(seed = 123)
contaminated <- Data_Contamination(sim_data, percentage = 5, seed = 1, type = "shift")
str(contaminated)

}
