% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM_ErrVar.R
\name{LMM.Error_variance}
\alias{LMM.Error_variance}
\title{Calculate Error Variance from a Linear Mixed Model (LMM)}
\usage{
LMM.Error_variance(data)
}
\arguments{
\item{data}{A data frame containing genotype-environment data.}
}
\value{
A numeric value representing the error variance.
}
\description{
This function computes the error variance using a linear mixed model (LMM) for a given dataset.
Depending on whether there is a single environment or a single genotype, the model is adjusted accordingly.
}
\examples{
sim_data <- sim.amb(seed = 123)
rep_data <- Create.Replications(sim_data, reps = 2, sig = 1)
err_var <- LMM.Error_variance(rep_data)
err_var

}
