% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsapply.R
\name{Metricsapply}
\alias{Metricsapply}
\title{Apply a Metric to a List of Data and Optionally Test Hypotheses}
\usage{
Metricsapply(
  list,
  metric,
  hypothesis.testing = FALSE,
  h0 = 100,
  side = "greater"
)
}
\arguments{
\item{list}{A list of data elements.}

\item{metric}{A character string specifying the metric to extract from each element.}

\item{hypothesis.testing}{Logical; if \code{TRUE}, performs a t-test for each metric against \code{h0}. Default is \code{FALSE}.}

\item{h0}{A numeric value for the null hypothesis mean in t-tests. Default is 100.}

\item{side}{A character string specifying the alternative hypothesis for t-tests (\code{"greater"} or \code{"less"}). Default is \code{"greater"}.}
}
\value{
A list containing the calculated metric values and, if requested, the corresponding p-values.
}
\description{
This function applies a specified metric to each element in a list of data,
and can optionally perform hypothesis testing on the metric values.
}
\examples{
# Assuming a list of simulated metric outputs:
metric_list <- list(model1 = list(Metric = c(101, 102)),
                    model2 = list(Metric = c(98, 99)))
result <- Metricsapply(metric_list, metric = "Metric", hypothesis.testing = TRUE, h0 = 100)
result

}
