% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Weights.R
\name{R_Weights}
\alias{R_Weights}
\title{Calculate Model Weights for LMM and RLMM}
\usage{
R_Weights(data, errorvariances)
}
\arguments{
\item{data}{A data frame containing genotype-environment data.}

\item{errorvariances}{A list containing error variances for genotypes and environments,
as computed by \code{Error_Var()}.}
}
\value{
A list containing sub-lists for LMM and RLMM weights with several weighting schemes.
}
\description{
This function calculates various weighting schemes for the data based on error variances
obtained from LMM and RLMM analyses, as well as weights derived from robust linear models.
}
\examples{
sim_data <- sim_amb(seed = 123)
rep_data <- Create_Replications(sim_data, reps = 2, sig = 1)
err_vars <- Error_Var(rep_data)
weights <- R_Weights(rep_data, errorvariances = err_vars)
names(weights)

}
