% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAMMI_Model.R
\name{rammi.model}
\alias{rammi.model}
\title{Fit a Robust AMMI (R-AMMI) Model}
\usage{
rammi.model(dataframe, weight = NULL, Ncomp = 2)
}
\arguments{
\item{dataframe}{A data frame with genotype-environment data.}

\item{weight}{A numeric vector of weights. If \code{NULL}, the regression is run without weights.}

\item{Ncomp}{Number of SVD components to retain. Default is 2.}
}
\value{
A list containing:
  \describe{
    \item{residual}{The residual matrix from the robust regression.}
    \item{matrix.fitted}{The matrix of fitted values from the robust regression.}
    \item{reduced.SVD}{The robust SVD reconstruction using \code{Ncomp} components.}
    \item{SVD}{The robust SVD object.}
  }
}
\description{
This function fits a robust AMMI model using robust linear regression.
It applies the regression to median-transformed data and then performs a robust SVD on the residuals.
}
\examples{
sim_data <- sim.amb(seed = 123)
rammi_res <- rammi.model(sim_data, weight = rep(1, nrow(sim_data)), Ncomp = 2)
names(rammi_res)

}
