% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLMM_ErrVar.R
\name{RLMM.Error_variance}
\alias{RLMM.Error_variance}
\title{Calculate Error Variance from a Robust Linear Mixed Model (RLMM)}
\usage{
RLMM.Error_variance(data)
}
\arguments{
\item{data}{A data frame containing genotype-environment data.}
}
\value{
A numeric value representing the error variance.
}
\description{
This function computes the error variance using a robust linear mixed model (RLMM)
for a given dataset. The model adjusts based on whether there is a single environment or genotype.
}
\examples{
sim_data <- sim.amb(seed = 123)
rep_data <- Create.Replications(sim_data, reps = 2, sig = 1)
err_var <- RLMM.Error_variance(rep_data)
err_var

}
