% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWAMMI_Model.R
\name{rwammi.model}
\alias{rwammi.model}
\title{Fit a Robust Weighted AMMI (RW-AMMI) Model}
\usage{
rwammi.model(dataframe, weight = NULL, Ncomp = 2, rSVD = FALSE)
}
\arguments{
\item{dataframe}{A data frame with genotype-environment data.}

\item{weight}{A numeric vector of weights. Must be provided.}

\item{Ncomp}{Number of SVD components to retain. Default is 2.}

\item{rSVD}{Logical; if \code{TRUE}, performs robust SVD. Default is \code{FALSE}.}
}
\value{
A list containing:
  \describe{
    \item{residual}{The residual matrix from the robust weighted regression.}
    \item{matrix.fitted}{The matrix of fitted values from the regression.}
    \item{reduced.SVD}{The weighted SVD reconstruction.}
    \item{SVD}{The SVD object of the reduced matrix.}
  }
}
\description{
This function fits a robust weighted AMMI model by applying robust linear regression
with weights and then performing a weighted (or robust) SVD on the residuals.
}
\examples{
sim_data <- sim.amb(seed = 123)
rwammi_res <- rwammi.model(sim_data, weight = rep(1, nrow(sim_data)), Ncomp = 2, rSVD = FALSE)
names(rwammi_res)

}
