% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_usable_data}
\alias{transform_usable_data}
\title{Transform Data by Applying a Function}
\usage{
transform_usable_data(dataframe, func, type = c("dataframe", "matrix"))
}
\arguments{
\item{dataframe}{A data frame containing genotype-environment data.}

\item{func}{A function to apply on the \code{yield} column (e.g., \code{mean} or \code{median}).}

\item{type}{A character string specifying the output type: either \code{"dataframe"} or \code{"matrix"}.
Default is \code{"dataframe"}.}
}
\value{
Either a summarized data frame or a matrix of transformed yield values.
}
\description{
This function transforms genotype-environment data either into a summarized data frame
or into a matrix by applying a specified function (e.g., \code{mean} or \code{median})
on the yield values grouped by genotype and environment.
}
\examples{
sim_data <- sim_amb(seed = 123)
trans_df <- transform_usable_data(sim_data, mean, type = "dataframe")
str(trans_df)

}
