% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAMMI_Model.R
\name{wammi_model}
\alias{wammi_model}
\title{Fit a Weighted AMMI (W-AMMI) Model}
\usage{
wammi_model(data, weight = NULL, Ncomp = 2)
}
\arguments{
\item{data}{A data frame with genotype-environment data.}

\item{weight}{A numeric vector of weights. Must be provided.}

\item{Ncomp}{Number of SVD components to retain. Default is 2.}
}
\value{
A list containing:
  \describe{
    \item{dataframe}{The transformed data with an added column of weighted residuals.}
    \item{residual}{The residual matrix from the weighted regression.}
    \item{matrix.fitted}{The matrix of fitted values from the weighted regression.}
    \item{reduced.SVD}{The weighted SVD reconstruction.}
    \item{SVD}{The SVD object of the weighted residual matrix.}
  }
}
\description{
This function fits a weighted AMMI model by performing weighted linear regression
and then applying a weighted SVD on the residuals.
}
\examples{
sim_data <- sim.amb(seed = 123)
# Assume weight vector is defined appropriately
weight_vec <- rep(1, nrow(sim_data))
wammi_res <- wammi.model(sim_data, weight = weight_vec, Ncomp = 2)
names(wammi_res)

}
