% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weighted_SVD.R
\name{WeightedSvdRes}
\alias{WeightedSvdRes}
\title{Perform Weighted (or Robust) Singular Value Decomposition}
\usage{
WeightedSvdRes(data, weight = NULL, Ncomp = 2, robust = FALSE)
}
\arguments{
\item{data}{A numeric matrix representing the data.}

\item{weight}{A numeric vector representing weights for the SVD. Default is \code{NULL}.}

\item{Ncomp}{The number of components to retain. Default is 2.}

\item{robust}{Logical; if \code{TRUE}, performs robust SVD using \code{rSVDdpd}. Default is \code{FALSE}.}
}
\value{
A matrix representing the reconstructed data from the weighted (or robust) SVD.
}
\description{
This function performs a weighted singular value decomposition (SVD) on a data matrix.
Optionally, a robust SVD method (\code{rSVDdpd}) can be used.
}
\examples{
# Create a random matrix and perform weighted SVD with 2 components
mat <- matrix(rnorm(100), nrow = 10)
W <- rep(1, length(mat))
svd_res <- WeightedSvdRes(mat, weight = W, Ncomp = 2)
dim(svd_res)

}
